requirements.txt
accelerate==0.31.0
aiohttp==3.9.5
aiosignal==1.3.1
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.4.0
asttokens==2.4.1
async-timeout==4.0.3
attrs==23.2.0
backcall==0.2.0
beautifulsoup4==4.12.3
bitsandbytes==0.42.0
blinker==1.8.2
certifi==2024.6.2
charset-normalizer==3.3.2
clarifai==10.5.0
clarifai-grpc==10.5.1
click==8.1.7
cloudpickle==3.0.0
comm==0.2.2
contextlib2==21.6.0
contourpy==1.2.1
cycler==0.12.1
dataclasses-json==0.5.14
debugpy==1.6.7
decorator==4.4.2
dill==0.3.8
distro==1.9.0
docker-pycreds==0.4.0
docopt==0.6.2
entrypoints==0.4
exceptiongroup==1.2.1
executing==2.0.1
Farama-Notifications==0.0.4
filelock==3.14.0
Flask==3.0.3
fonttools==4.53.0
frozenlist==1.4.1
fsspec==2024.6.0
gitdb==4.0.11
GitPython==3.1.43
googleapis-common-protos==1.63.1
greenlet==3.0.3
groq==0.9.0
grpcio==1.64.1
gym==0.26.2
gym-notices==0.0.8
gymnasium==0.29.1
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
huggingface-hub==0.23.3
hydra-core==1.3.2
idna==3.7
imageio==2.34.1
imageio-ffmpeg==0.5.1
importlib_metadata==7.1.0
importlib_resources==6.4.0
inquirerpy==0.3.4
ipykernel==6.29.5
ipython==8.12.0
ipywidgets==8.1.3
itsdangerous==2.2.0
jedi==0.19.1
Jinja2==3.1.4
joblib==1.4.2
Js2Py==0.74
jupyter-client==7.3.4
jupyter_core==5.7.2
jupyterlab_widgets==3.0.11
kiwisolver==1.4.5
langchain==0.0.225
langchainplus-sdk==0.0.20
llvmlite==0.43.0
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.21.3
matplotlib==3.9.0
matplotlib-inline==0.1.7
mdurl==0.1.2
moviepy==1.0.3
mpmath==1.3.0
multidict==6.0.5
mypy-extensions==1.0.0
nest_asyncio==1.6.0
networkx==3.2.1
numba==0.60.0
numexpr==2.10.0
numpy==1.26.4
numpy-groupies==0.11.1
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.18.1
nvidia-nvjitlink-cu12==12.5.40
nvidia-nvtx-cu12==12.1.105
ollama==0.2.1
omegaconf==2.3.0
openai==1.5.0
openapi-schema-pydantic==1.2.4
opencv-python==4.10.0.82
-e git+https://github.com/HumanCompatibleAI/overcooked_ai.git@cb2e50cae95accbe4618879d88e565c87c54b1c3#egg=overcooked_ai&subdirectory=../../overcooked_ai
packaging==24.1
pandas==2.2.2
parso==0.8.4
pettingzoo==1.24.3
pexpect==4.9.0
pfzy==0.3.4
pickleshare==0.7.5
pillow==10.3.0
pip==24.0
pipwin==0.5.2
platformdirs==4.2.2
plotly==5.22.0
proglog==0.1.10
prompt_toolkit==3.0.47
protobuf==5.27.1
psutil==5.9.0
ptyprocess==0.7.0
pure-eval==0.2.2
pydantic==1.10.16
pydantic_core==2.18.4
pygame==2.5.2
pygame-ce==2.5.0
pygame-gui==0.6.9
Pygments==2.18.0
pyjsparser==2.7.1
pyparsing==3.1.2
PyPrind==2.11.3
pySmartDL==1.3.4
python-dateutil==2.9.0
python-dotenv==1.0.1
python-i18n==0.3.9
python-rapidjson==1.17
pytz==2024.1
PyYAML==6.0.1
pyzmq==25.1.2
regex==2024.5.15
requests==2.32.3
ribs @ git+https://github.com/icaros-usc/pyribs.git@c2a77a3e8cb4011d7741d9105128c0aed99a108e
rich==13.7.1
safetensors==0.4.3
schema==0.7.5
scikit-learn==1.5.0
scipy==1.13.1
sentry-sdk==2.5.1
setproctitle==1.3.3
setuptools==69.5.1
shapely==2.0.4
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.5
SQLAlchemy==2.0.24
stack-data==0.6.2
sympy==1.12.1
tabulate==0.9.0
tenacity==8.3.0
threadpoolctl==3.5.0
tokenizers==0.19.1
torch==2.1.2
tornado==6.1
tqdm==4.66.4
traitlets==5.14.3
transformers==4.41.2
triton==2.1.0
tritonclient==2.46.0
typing_extensions==4.12.2
typing-inspect==0.9.0
tzdata==2024.1
tzlocal==5.2
urllib3==2.2.1
wandb==0.17.1
wcwidth==0.2.13
Werkzeug==3.0.3
wheel==0.43.0
widgetsnbextension==4.0.11
yarl==1.9.4
zipp==3.19.2

